fix: accept fractional hours in monitor-stats API
Changed int() to float() for the hours param so 0.167 (10 min)
and 0.5 (30 min) work.

+++ skills/baby-monitor/dashboard/app.py
@@ -645,7 +645,7 @@ def api_monitor_stats():
     """
     from collections import Counter
 
-    hours = int(request.args.get("hours", 24))
+    hours = float(request.args.get("hours", 24))
     cutoff = datetime.now(timezone.utc) - timedelta(hours=hours)
 
     entries = []
